Use .env for environment variables
